Rename payment_methods to transaction_providers
